move test to codegen_test
